Remove old settings screen
